Stats par antenne et par region
